requirements.txt
pandas
requests
python-telegram-bot==20.3
